+++ src/engine/engine_collision_driver.c
@@ -118,8 +118,7 @@ int mj_maxContact(const mjModel* m, int g1, int g2, int has_margin) {
     return 1;
   }
 
-  if (type1 == mjGEOM_CAPSULE || type2 == mjGEOM_CAPSULE ||
-      type1 == mjGEOM_CYLINDER || type2 == mjGEOM_CYLINDER) {
+  if (type1 == mjGEOM_CAPSULE || type2 == mjGEOM_CAPSULE) {
     return 5;
   }
 
@@ -151,7 +150,7 @@ int mj_maxContact(const mjModel* m, int g1, int g2, int has_margin) {
     }
   }
 
-  // 4 contacts for mesh-mesh or mesh-box without margins, 5 with margins
+  // 4 contacts for box, cylinder, and mesh collisions without margins, 5 with margins
   return has_margin ? 5 : 4;
 }