+++ test/engine/engine_collision_driver_test.cc
@@ -469,9 +469,9 @@ TEST_F(MjCollisionTest, MaxContact) {
   EXPECT_EQ(mj_maxContact(m.get(), ellipsoid, capsule, -1), 1);
   EXPECT_EQ(mj_maxContact(m.get(), capsule, cylinder, -1), 5);
   EXPECT_EQ(mj_maxContact(m.get(), capsule, mesh, -1), 5);
-  EXPECT_EQ(mj_maxContact(m.get(), cylinder, cylinder, -1), 5);
-  EXPECT_EQ(mj_maxContact(m.get(), cylinder, box, -1), 5);
-  EXPECT_EQ(mj_maxContact(m.get(), cylinder, mesh, -1), 5);
+  EXPECT_EQ(mj_maxContact(m.get(), cylinder, cylinder, -1), 4);
+  EXPECT_EQ(mj_maxContact(m.get(), cylinder, box, -1), 4);
+  EXPECT_EQ(mj_maxContact(m.get(), cylinder, mesh, -1), 4);
 }
 
 }  // namespace